+++ ios/Flutter/flutter_export_environment.sh
@@ -1,0 +1,10 @@
+#!/bin/sh
+# This is a generated file; do not edit or check into version control.
+export "FLUTTER_ROOT=E:\Programs\flutter"
+export "FLUTTER_APPLICATION_PATH=C:\Users\dartz\AndroidStudioProjects\amiibo_network"
+export "FLUTTER_TARGET=lib\main.dart"
+export "FLUTTER_BUILD_DIR=build"
+export "SYMROOT=${SOURCE_ROOT}/../build\ios"
+export "FLUTTER_FRAMEWORK_DIR=E:\Programs\flutter\bin\cache\artifacts\engine\ios"
+export "FLUTTER_BUILD_NAME=1.2.2"
+export "FLUTTER_BUILD_NUMBER=27"