+++ lib/service/notification_service.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/services.dart';
+
+class NotificationService {
+  static const _channel = const MethodChannel("com.dartz.amiibo_network/notification");
+
+  static Future<bool> sendNotification(Map<String,dynamic> args) async {
+    //TODO IOS Platform Channel
+    try {
+      final bool result = await _channel.invokeMethod('notification', args);
+      return result;
+    } on PlatformException catch (e) {
+      print(e);
+      return false;
+    }
+  }
+
+}