+++ lib/service/storage.dart
@@ -6,11 +6,11 @@ import 'package:path_provider/path_provider.dart';
 bool checkPermission(PermissionStatus permission) {
   switch(permission){
     case(PermissionStatus.granted):
-    case(PermissionStatus.unknown):
       return true;
     case(PermissionStatus.denied):
     case(PermissionStatus.neverAskAgain):
     case(PermissionStatus.restricted):
+    case(PermissionStatus.unknown):
     default:
       return false;
   }
@@ -18,33 +18,25 @@ bool checkPermission(PermissionStatus permission) {
 
 Future<File> createFile([String name = 'MyAmiiboNetwork', String type = 'json']) async{
   Directory dir;
-  if(Platform.isAndroid) dir = await getExternalStorageDirectory();
+  StorageDirectory storage = StorageDirectory.documents;
+  if(type == 'png') storage = StorageDirectory.pictures;
+  if(Platform.isAndroid) dir = (await getExternalStorageDirectories(type: storage)).first;
   else dir = await getApplicationDocumentsDirectory();
   String path = '${dir.path}/${name}_${DateTime.now().toString().substring(0,10)}.$type';
   final File file = File(path);
   return file;
 }
 
-Future<File> createCacheFile(String path) async{
-  final File file = File(path);
-  final Directory dir = await getTemporaryDirectory();
-  final File copyFile = await file.copy('${dir.path}/cacheAmiibo.json');
-  return copyFile;
-}
-
-Map<String,dynamic> readFile1(File file) {
-  String data = file.readAsStringSync();
-  final Map<String, dynamic> jResult = jsonDecode(data);
-  if(!jResult.containsKey('amiibo')) return null;
-  else return jResult;
-}
-
-Map<String,dynamic> readFile(String path) {
-  final file = File(path);
-  String data = file.readAsStringSync();
-  final Map<String, dynamic> jResult = jsonDecode(data);
-  if(!jResult.containsKey('amiibo')) return null;
-  else return jResult;
+Map<String,dynamic> readFile(File file) {
+  try{
+    String data = file.readAsStringSync();
+    final Map<String, dynamic> jResult = jsonDecode(data);
+    if(!jResult.containsKey('amiibo')) return null;
+    else return jResult;
+  }catch(e){
+    print(e.toString());
+    return null;
+  }
 }
 
 void writeFile(Map<String,dynamic> arg) =>