+++ lib/utils/urls_constants.dart
@@ -1,3 +1,4 @@
 const String privacyPolicy = 'https://github.com/EdwynZN/amiibo_network/blob/master/Privacy%20Policy.md';
 const String github = 'https://github.com/EdwynZN/amiibo_network';
 const String reportIssue = 'https://github.com/EdwynZN/amiibo_network/issues';
+const String kofi = 'https://ko-fi.com/hcgdartz';